Add EditForm component, add editAdRequest to adsReducer, update FormPattern

+++ controllers/ads.controller.js
@@ -43,30 +43,32 @@ exports.getOne = async (req, res) => {
 }
 
 exports.addNew = async(req, res) => {
+    let filePath = null;
     try {
         sanitize(req.body);
         const {title, content, publicationDate, price, location, author} = req.body;
         const image = req.file;
-        const filePath = image ? path.join(process.cwd(), 'public', 'uploads', image.filename) : null;
+        if(image) filePath = path.join(process.cwd(), 'public', 'uploads', image.filename);
         if(title && content && publicationDate && image && price && location && author){
             const imageMimeTypes = ['image/jpeg', 'image/png', 'image/gif'];
             const fileType = await getImageFileType(image);
             if(imageMimeTypes.includes(fileType)){
                 const newAd = new Ad({ title, content, publicationDate, image: image.filename, price, location, author});
                 await newAd.save();
-                return res.json({ message: 'OK' });
+                const newAdWithAuthor = await Ad.findById(newAd._id).populate('author');
+                return res.json({ message: newAdWithAuthor });
             } else {
-                await removeFile(filePath);
+                if(filePath)await removeFile(filePath);
                 return res.status(400).json({ message: 'Invalid file'});
             }
         } else {
-            await removeFile(filePath);
+            if(filePath)await removeFile(filePath);
             return res.status(400).json({ message: 'All params required' });
         }
     }
     catch(error){
         if (error.name === 'ValidationError') {
-            await removeFile(filePath);
+            if(filePath) await removeFile(filePath);
             return res.status(400).json({
                 message: 'Invalid params',
                 errors: Object.values(error.errors).map(err => err.message)
@@ -77,14 +79,16 @@ exports.addNew = async(req, res) => {
 };
 
 exports.editOne = async(req, res) => {
+    let filePath = null;
     try {
         sanitize(req.body);
-        const {title, content, publicationDate, price, location, author} = req.body;
+        const {title, content, publicationDate, price, location} = req.body;
         const image = req.file;
+        const authorId = typeof req.body.author === 'object' ? req.body.author._id : req.body.author;
         const adToEdit = await Ad.findById(req.params.id);
-        const filePath = image ? path.join(process.cwd(), 'public', 'uploads', image.filename) : null;
+        filePath = image ? path.join(process.cwd(), 'public', 'uploads', image.filename) : null;
         if(adToEdit){
-            if(title && content && publicationDate && price && location && author){
+            if(title && content && publicationDate && price && location && authorId){
                 let imageFilename = null;
                 if(image) {
                     const oldFilePath = path.join(process.cwd(), 'public', 'uploads', adToEdit.image);
@@ -94,31 +98,34 @@ exports.editOne = async(req, res) => {
                         imageFilename = image.filename;
                         await removeFile(oldFilePath);
                     } else {
-                        await removeFile(filePath);
+                        if(filePath) await removeFile(filePath);
                         return res.status(400).json({ message: 'Invalid file'});
                     } 
                 } else imageFilename = adToEdit.image;
-                await adToEdit.updateOne({ $set: { title, content, publicationDate, image: imageFilename, price, location, author}});              
-                return res.json({ message: 'OK' });
+                const authorObjectId = new mongoose.Types.ObjectId(authorId);
+                await adToEdit.updateOne({ $set: { title, content, publicationDate, image: imageFilename, price, location, author: authorObjectId } });
+                const updatedAd = await Ad.findById(adToEdit._id).populate('author');  
+                console.log('updatedAd', updatedAd)        ;
+                return res.json({ message: updatedAd });
             } else {
-                await removeFile(filePath);
+                if(filePath) await removeFile(filePath);
                 return res.status(400).json({ message: 'All params required' });
             }
         } else {
-            await removeFile(filePath);
+            if(filePath) await removeFile(filePath);
             return res.status(404).json({ message: 'Not found' });
         }
     }
     catch(error){
         if (error.name === 'ValidationError') {
-            await removeFile(filePath);
+            if(filePath) await removeFile(filePath);
             return res.status(400).json({
                 message: 'Invalid params',
                 errors: Object.values(error.errors).map(err => err.message)
             });
         }
-        await removeFile(filePath);
-        return res.status(500).json({ message: 'Internal Server Error' });
+        if(filePath) await removeFile(filePath);
+        return res.status(500).json({ message: error });
     }
 };
